+++ src/main/java/com/shop/cart/repo/CartRepository.java
@@ -12,6 +12,6 @@ import com.shop.cart.model.Cart;
  */
  
 @Repository
-public interface CartRepository extends CrudRepository<Cart, Integer>{
+public interface CartRepository extends CrudRepository<Cart, Long>{
  
 }
